move cpp headers into +private.h
