+++ ios/browser/url_sanitizer/url_sanitizer_service_factory+private.h
@@ -1,0 +1,55 @@
+// Copyright (c) 2023 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
+#ifndef BRAVE_IOS_BROWSER_URL_SANITIZER_URL_SANITIZER_SERVICE_FACTORY_PRIVATE_H_
+#define BRAVE_IOS_BROWSER_URL_SANITIZER_URL_SANITIZER_SERVICE_FACTORY_PRIVATE_H_
+
+#include <memory>
+
+#include "brave/components/url_sanitizer/browser/url_sanitizer_service.h"
+#include "brave/ios/browser/keyed_service/keyed_service_factory_wrapper.h"
+#include "components/keyed_service/ios/browser_state_keyed_service_factory.h"
+
+namespace base {
+template <typename T>
+class NoDestructor;
+}  // namespace base
+
+namespace web {
+class BrowserState;
+}  // namespace web
+
+class KeyedService;
+class ChromeBrowserState;
+
+namespace brave {
+
+class URLSanitizerServiceFactory : public BrowserStateKeyedServiceFactory {
+ public:
+  static brave::URLSanitizerService* GetServiceForState(
+      ChromeBrowserState* browser_state);
+
+  static URLSanitizerServiceFactory* GetInstance();
+
+ protected:
+  bool ServiceIsCreatedWithBrowserState() const override;
+
+ private:
+  friend base::NoDestructor<URLSanitizerServiceFactory>;
+
+  URLSanitizerServiceFactory();
+  ~URLSanitizerServiceFactory() override;
+
+  // BrowserStateKeyedServiceFactory implementation.
+  std::unique_ptr<KeyedService> BuildServiceInstanceFor(
+      web::BrowserState* context) const override;
+  bool ServiceIsNULLWhileTesting() const override;
+  web::BrowserState* GetBrowserStateToUse(
+      web::BrowserState* context) const override;
+};
+
+}  // namespace brave
+
+#endif  // BRAVE_IOS_BROWSER_URL_SANITIZER_URL_SANITIZER_SERVICE_FACTORY_PRIVATE_H_
